requirements.txt
# AI Secretary System - Dependencies
# Python 3.11+ required

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Task Queue
celery>=5.3.0
redis>=5.0.0

# AI/LLM
langgraph>=0.0.40
langchain>=0.1.0
langchain-anthropic>=0.1.0
anthropic>=0.18.0

# Browser Automation
playwright>=1.40.0

# Database
supabase>=2.0.0

# Email Integration
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0

# LINE Integration
line-bot-sdk>=3.5.0

# Security
cryptography>=41.0.0
python-jose[cryptography]>=3.3.0

# Utilities
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
httpx>=0.25.0

